[Feature #164361854]Retructures tests
